Support for `--include` that works the opposite as `--ignore`
Useful if you need to process a directory that contains a lot of other things.
Instead of ignoreing stuff you can now filter out just the files you need.

This also changes `--schema` and `--ignore` to support comma seperated arguments
instead of the argument having to be repeated. This helps cut down the amount of
processing needed to generate the argument list in some cases.